chore: Move BuyTicketsModal into BuyTicketsModal directory
