+++ src/views/LotteryV2/components/BuyTicketsButton.tsx
@@ -3,7 +3,7 @@ import { Button, useModal, WaitIcon, ButtonProps } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { useLottery } from 'state/hooks'
 import { LotteryStatus } from 'config/constants/types'
-import BuyTicketsModal from './BuyTicketsModal'
+import BuyTicketsModal from './BuyTicketsModal/BuyTicketsModal'
 
 interface BuyTicketsButtonProps extends ButtonProps {
   disabled?: boolean
