+++ src/views/LotteryV2/components/BuyTicketsModal/TicketInput.tsx
