--- src/views/LotteryV2/components/EditNumbersModal/index.tsx
@@ -1,2 +1,0 @@
-export { default as EditNumbersModal } from './EditNumbersModal'
-export * from './useTicketsReducer'
